+++ spark_api/src/main.py
@@ -1,0 +1,16 @@
+from concurrent import futures
+
+import grpc
+from sparkapi_pb2_grpc import SparkApiServicer, add_SparkApiServicer_to_server
+
+
+def serve():
+    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
+    add_SparkApiServicer_to_server(SparkApiServicer(), server)
+    server.add_insecure_port('[::]:50051')
+    server.start()
+    server.wait_for_termination()
+
+
+if __name__ == '__main__':
+    serve()
